# server/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Vello Trade API
  description: |
    REST API for the Vello Trade B2B marketplace mobile app (Expo). JSON routes are mounted under **`/api`** (the mobile client uses `baseURL: {host}/api`).

    **Socket.IO** runs on the same HTTP server. After login, connect with JWT in handshake `auth: { token: "<jwt>" }`.
    Events include `room:join`, `chat:typing`, `message:new`, `chat:room:updated`. See the repository README for details.
  version: 1.0.0
servers:
  - url: http://localhost:4000
    description: Local development
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - bearerAuth: []
paths:
...
  /api/products:
    get:
      tags: [Products]
      summary: List products (global catalog + listings; optional filters)
      security: []
      parameters:
        - in: query
          name: q
          schema:
            type: string
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: companyId
          schema:
            type: string
        - in: query
          name: sellerId
          schema:
            type: string
    post:
      tags: [Products]
      summary: Create listing (company / distributor)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, category, price, quantity]
              properties:
                name:
                  type: string
                category:
                  type: string
                price:
                  type: number
                quantity:
                  type: integer
                description:
                  type: string
                images:
                  type: array
                  items:
                    type: string
  /api/products/{id}:
    get:
      tags: [Products]
      summary: Product detail
      security: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
    patch:
      tags: [Products]
      summary: Update own product
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
    delete:
      tags: [Products]
      summary: Delete own product
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
...
  /api/users/retail:
    post:
      tags: [Users]
      summary: Distributor creates retail user
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [password]
              properties:
                email:
                  type: string
                phone:
                  type: string
                password:
                  type: string
                name:
                  type: string
                businessName:
                  type: string
...
  /api/users/link-distributor:
    patch:
      tags: [Users]
      summary: Attach distributor to company
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [companyId]
              properties:
                companyId:
                  type: string
  /api/users/link-retail:
    patch:
      tags: [Users]
      summary: Attach retail to distributor
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [distributorId]
              properties:
                distributorId:
                  type: string
  /api/payments:
    get:
      tags: [Payments]
      summary: My payments
    post:
      tags: [Payments]
      summary: Create payment record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [payeeId, amount]
              properties:
                payeeId:
                  type: string
                amount:
                  type: number
                method:
                  type: string
                currency:
                  type: string
                orderId:
                  type: string
                invoiceId:
                  type: string
                notes:
                  type: string
  /api/payments/{id}/status:
    patch:
      tags: [Payments]
      summary: Update payment status
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [pending, completed, failed, refunded]
...
  /api/admin/catalog/products/{id}:
    patch:
      tags: [Admin]
      summary: Update global SKU
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
    delete:
      tags: [Admin]
      summary: Delete global SKU
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
...